+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 Unreleased
 ----------
 
+* BREAKING CHANGE: Redesign IconField. Instead of ``IconField("envelope")`` call it with the
+  full icon class e.g. ``IconField("fa fa-envelope")``.
 * Drop support for Django 3.1
 * Drop support for Python 3.6